CompilationVisitor as an inner class of Compilation
